load test and split files
