+++ .github/workflows/go_integration_tests.yml
@@ -21,4 +21,4 @@ jobs:
         uses: actions/checkout@v2
         
       - name: Run the integration test
-        run: go test -timeout 7m -run TestIntegration ./integration/...
+        run: go test -timeout 15m -run TestIntegration ./integration/...
