+++ integration/dvotingdela.go
@@ -109,7 +109,7 @@ type dVotingNode struct {
 
 // Creates n dela nodes using tempDir as root to file path and returns an array
 // of nodes or error
-func setupDVotingNodes(t require.TestingT, numberOfNodes int, tempDir string, node dVotingCosiDela) []dVotingCosiDela {
+func setupDVotingNodes(t require.TestingT, numberOfNodes int, tempDir string, nodeSetup dVotingCosiDela) []dVotingCosiDela {
 
 	wait := sync.WaitGroup{}
 
@@ -137,8 +137,9 @@ func setupDVotingNodes(t require.TestingT, numberOfNodes int, tempDir string, no
 		delaNodes = append(delaNodes, node)
 		dVotingNodes = append(dVotingNodes, node)
 	}
-	if node != nil {
-		node.Setup(delaNodes...)
+
+	if nodeSetup != nil {
+		nodeSetup.Setup(delaNodes...)
 	} else {
 		delaNodes[0].Setup(delaNodes[1:]...)
 	}
